feat: Implement OpenAI client and validation utilities
- Added OpenAI client initialization with API key validation.
- Created validation utility for OpenAI chat completions using Zod schema.
- Developed complete post analysis functionality, integrating title generation, categorization, and sentiment analysis.
- Implemented Redis post group management with title and sentiment summary generation.
- Added Redis client initialization and connection handling with error management.
- Created deduplication listener for posts using cosine similarity on embeddings.
- Seeded Redis database with mock data and provided command-line interface for seeding operations.
- Added embedding retrieval with retry logic for OpenAI API calls.
