+++ src/analysis/generateTitle.ts
@@ -1,6 +1,9 @@
+// This file has been moved to the analysis subfolder.
+// Please update your imports accordingly.
+import { generateTitleForPost } from './generateTitle.js';
 import type OpenAI from 'openai';
-import openai from './openaiClient.js';
-import { callOpenAIWithValidation } from './openaiValidationUtil.js';
+import openai from '../openai/openaiClient.js';
+import { callOpenAIWithValidation } from '../openai/openaiValidationUtil.js';
 import { z } from 'zod';
 
 // Title generation result type
